requirements.txt
aiohttp>=3.9.0
dnspython>=2.4.0
python-whois>=0.9.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
rich>=13.7.0
jinja2>=3.1.0
networkx>=3.2.0
pyvis>=0.3.2
cryptography>=42.0.0
mmh3>=4.0.0
pyyaml>=6.0.0
pydantic>=2.7.0
click>=8.1.0
aiofiles>=23.2.0
pytest>=7.4.0
pytest-asyncio>=0.23.0
